add code to html and js to allow decimal numbers and update the readme

+++ js/scripts.js
@@ -1,8 +1,8 @@
 $(document).ready(function() {
   $("form#triform").submit(function(event) {
-    const sideA = parseInt($('input#sideA').val());
-    const sideB = parseInt($('input#sideB').val());
-    const sideC = parseInt($('input#sideC').val());
+    const sideA = parseFloat($('input#sideA').val());
+    const sideB = parseFloat($('input#sideB').val());
+    const sideC = parseFloat($('input#sideC').val());
     $(".isosceles").hide();
     $(".equilateral").hide();
     $(".scalene").hide();
